fix(examples): keep App Studio client-safe and publish the Grok ACP port
Stop the browser from importing sandbox code. Publish port 2419 so Grok Build can connect. Avoid a title-update render loop. Read preview URLs from assistant text.
